Implemented SnapshotTestCase and WebpageSnapshotTestCase to clean up snapshot test configuration.
