+++ Tests/AppTests/RSSTests.swift
@@ -4,13 +4,8 @@ import SnapshotTesting
 import XCTVapor
 
 
-class RSSTests: AppTestCase {
-    
-    override func setUpWithError() throws {
-        try super.setUpWithError()
-        SnapshotTesting.isRecording = false
-    }
-    
+class RSSTests: SnapshotTestCase {
+
     func test_render_item() throws {
         let item = RecentPackage(id: UUID(),
                                  repositoryOwner: "owner",
