+++ Tests/AppTests/SnapshotTestCase.swift
@@ -1,0 +1,11 @@
+import SnapshotTesting
+
+
+class SnapshotTestCase: AppTestCase {
+    override func setUpWithError() throws {
+        try super.setUpWithError()
+
+        SnapshotTesting.isRecording = false
+        SnapshotTesting.diffTool = "ksdiff"
+    }
+}
