+++ Tests/AppTests/WebpageSnapshotTestCase.swift
@@ -1,0 +1,18 @@
+@testable import App
+
+import XCTVapor
+
+
+class WebpageSnapshotTestCase: SnapshotTestCase {
+    override func setUpWithError() throws {
+        try super.setUpWithError()
+
+        try XCTSkipIf((Environment.get("SKIP_SNAPSHOTS") ?? "false") == "true")
+        Current.date = { Date(timeIntervalSince1970: 0) }
+        TempWebRoot.cleanup()
+    }
+
+    override class func setUp() {
+        TempWebRoot.setup()
+    }
+}
